split common and ui components
