+++ src/components/common/slider.tsx
