+++ src/components/index.ts
@@ -1,2 +1,6 @@
-export { default as Slider } from "./slider";
-export { default as Chip } from "./chip";
+export { default as Slider } from "./common/slider";
+export { default as Chip } from "./common/chip";
+
+export { default as ColorInput } from "./ui/input";
+export { default as ColorMode } from "./ui/mode";
+export { default as ColorView } from "./ui/view";
